requirements.txt
Django
djangorestframework
django-allauth
django-cors-headers
django-filter
django-environ
django-redis
django-crispy-forms
crispy-tailwind
django-widget-tweaks
django-celery-beat

celery
redis
gunicorn
whitenoise

psycopg2-binary
dj-database-url

cloudinary
django-cloudinary-storage

weasyprint
reportlab
qrcode
openpyxl
pandas
matplotlib
seaborn
scikit-learn
statsmodels

python-dotenv
requests
pytest